requirements.txt
Flask>=3.0.0
gunicorn>=21.0.0
argon2-cffi>=23.1.0
PyYAML>=6.0
mutagen>=1.47.0
APScheduler>=3.10.0
